fix: Move AggregationHook.h to velox/vector/ from velox/exec/
Fix #15449

`velox/dwio/common/SelectiveColumnReaderInternal.h` includes
`velox/exec/AggregationHook.h`. But `velox/exec/**/*.h` aren't
installed when `VELOX_ENABLE_EXEC=OFF`.

`velox/dwio/` should not depend on `velox/exec/` because
`VELOX_BUILD_MINIMAL_WITH_DWIO=ON` sets `VELOX_ENABLE_EXEC=OFF`.
(Nimble uses `VELOX_BUILD_MINIMAL_WITH_DWIO=ON`.)

If we move `AggregationHook.h` to `velox/vector/` from `velox/exec/`,
`velox/dwio/common/SelectiveColumnReaderInternal.h` can include
`AggregationHook.h` because `velox/vector/` is always enabled.
