+++ velox/dwio/common/SelectiveColumnReaderInternal.h
@@ -21,8 +21,8 @@
 #include "velox/dwio/common/DirectDecoder.h"
 #include "velox/dwio/common/SelectiveColumnReader.h"
 #include "velox/dwio/common/TypeUtils.h"
-#include "velox/exec/AggregationHook.h"
 #include "velox/type/Timestamp.h"
+#include "velox/vector/AggregationHook.h"
 #include "velox/vector/ConstantVector.h"
 #include "velox/vector/DictionaryVector.h"
 #include "velox/vector/FlatVector.h"
