+++ velox/functions/lib/aggregates/SimpleNumericAggregate.h
@@ -16,7 +16,7 @@
 #pragma once
 
 #include "velox/exec/Aggregate.h"
-#include "velox/exec/AggregationHook.h"
+#include "velox/vector/AggregationHook.h"
 #include "velox/vector/DecodedVector.h"
 #include "velox/vector/FlatVector.h"
 #include "velox/vector/LazyVector.h"
