+++ velox/functions/lib/aggregates/tests/SumTestBase.h
@@ -17,10 +17,10 @@
 #include "folly/CPortability.h"
 
 #include "velox/common/base/tests/GTestUtils.h"
-#include "velox/exec/AggregationHook.h"
 #include "velox/exec/tests/utils/AssertQueryBuilder.h"
 #include "velox/exec/tests/utils/PlanBuilder.h"
 #include "velox/functions/lib/aggregates/tests/utils/AggregationTestBase.h"
+#include "velox/vector/AggregationHook.h"
 
 namespace facebook::velox::functions::aggregate::test {
